requirements.txt
boto3
facebook-sdk==2.0.0
